+++ apps/seller/lib/domain/repository/product_repository.dart
@@ -1,9 +1,9 @@
 import 'package:core/error/result.dart';
 
-import '../entities/create_product_params.dart';
+import '../entity/create_product_params.dart';
 
 abstract class ProductRepository {
-  Future<Result<void>> createProduct(CreateProductParams params);
+  Future<Result<void>> createProduct(AddProduct params);
 
   Future<Result<String>> getOwnerStoreId();
 }
